requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
twilio==8.10.0
google-generativeai>=0.3.1
pydantic==2.5.0
python-multipart==0.0.6
jinja2==3.1.2
